backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.13
asyncpg>=0.29.0
redis[asyncio]>=5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.25.0
openai>=1.3.0
pgvector>=0.2.4
langchain>=0.1.0
